refactor(ops): extract callable wrapper, xorq compat, normalize, root models
Pull four more self-contained chunks out of _core.py:

- _callable.py (60 lines): _CallableWrapper / _ensure_wrapped /
  _infer_unnest. Tiny module, used by serialization & utils externally.
- _xorq_compat.py (127 lines): _ensure_xorq_table,
  _patch_xorq_sortkey_compat, _rebind_to_backend,
  _rebind_to_canonical_backend. Pure ibis↔xorq bridge logic.
- _normalize.py (105 lines): _normalize_to_name and
  _normalize_join_predicate, the two arg-coercion helpers used by
  the user-facing API.
- _root_models.py (456 lines): root-model traversal
  (_find_all_root_models et al) and the join-aware field-prefixing
  / _right column-rename machinery. SemanticTableOp / SemanticJoinOp
  are imported lazily inside each function to break the import
  cycle with _core.

_core.py is now 4,186 lines (down from 4,621 last commit, 5,330 at
start of branch). All extractions are pure moves; _core.py re-imports
the names back from each submodule so internal call sites are unchanged.

930 tests pass.
