+++ src/boring_semantic_layer/ops/_core.py
@@ -91,98 +91,7 @@ class _RenamedResolver:
         return getattr(self._table, mapped)
 
 
-def _normalize_to_name(arg: str | Deferred) -> str:
-    """Convert a string or simple ``_.name`` Deferred to a plain string name.
-
-    Accepts a plain string (returned as-is) or a Deferred whose resolver is a
-    simple attribute access on the top-level ``_`` variable (e.g. ``_.origin``).
-
-    Complex expressions like ``_.distance.sum()`` or ``_.a.b`` are rejected
-    with a ``TypeError``.
-    """
-    if isinstance(arg, str):
-        return arg
-
-    # Duck-type: works for both ibis and xorq Deferred objects
-    resolver = getattr(arg, "_resolver", None)
-    if resolver is None:
-        raise TypeError(
-            f"Expected a string name or Deferred expression (_.name), got {type(arg).__name__}"
-        )
-
-    obj = getattr(resolver, "obj", None)
-
-    # Try attribute access first (_.name -> Attr resolver with .name)
-    name_wrapper = getattr(resolver, "name", None)
-
-    # Fall back to getitem access (_["name"] -> Item resolver with .indexer)
-    if name_wrapper is None:
-        name_wrapper = getattr(resolver, "indexer", None)
-
-    if name_wrapper is None or obj is None:
-        raise TypeError(
-            f"Only simple Deferred expressions like _.name or _['name'] are supported "
-            f"as positional arguments, got: {arg!r}"
-        )
-
-    # Reject chained access like _.a.b (obj would itself have an .obj attr)
-    if getattr(obj, "obj", None) is not None:
-        raise TypeError(
-            f"Only simple Deferred expressions like _.name or _['name'] are supported "
-            f"as positional arguments, got: {arg!r}"
-        )
-
-    # Attr.name / Item.indexer is a Just wrapper; unwrap via .value
-    raw_name = getattr(name_wrapper, "value", name_wrapper)
-    if not isinstance(raw_name, str):
-        raise TypeError(f"Could not extract string name from Deferred expression: {arg!r}")
-
-    return raw_name
-
-
-def _normalize_join_predicate(on):
-    """Normalize a join predicate to a two-argument callable.
-
-    Accepts:
-    - ``str`` – equi-join on a column present in both sides
-    - ``Deferred`` (``_.col``) – same, after extracting the name
-    - ``list[str | Deferred]`` – compound equi-join on multiple columns
-    - ``callable`` (non-Deferred) – returned as-is (existing lambda API)
-    - ``None`` – returned as-is (for cross joins)
-    """
-    if on is None:
-        return on
-
-    if isinstance(on, str):
-        name = on
-        return lambda left, right: getattr(left, name) == getattr(right, name)
-
-    if _is_deferred(on):
-        name = _normalize_to_name(on)
-        return lambda left, right: getattr(left, name) == getattr(right, name)
-
-    if isinstance(on, (list, tuple)):
-        names = [_normalize_to_name(item) for item in on]
-        if len(names) == 1:
-            name = names[0]
-            return lambda left, right: getattr(left, name) == getattr(right, name)
-
-        def _compound_predicate(left, right):
-            from functools import reduce
-            from operator import and_
-
-            preds = [getattr(left, n) == getattr(right, n) for n in names]
-            return reduce(and_, preds)
-
-        return _compound_predicate
-
-    if callable(on):
-        return on
-
-    raise TypeError(
-        f"join `on` must be a string, Deferred (_.col), list of strings/Deferred, "
-        f"or a callable, got {type(on).__name__}"
-    )
+from ._normalize import _normalize_join_predicate, _normalize_to_name  # noqa: E402, F401
 
 
 if TYPE_CHECKING:
@@ -195,117 +104,12 @@ if TYPE_CHECKING:
     )
 
 
-def _patch_xorq_sortkey_compat():
-    """Register a map_ibis handler so ibis SortKey → xorq SortKey.
-
-    ibis 11 uses ``SortKey.expr``, ibis 12 renamed it to ``SortKey.arg``,
-    while xorq's vendored ibis keeps ``SortKey.expr``.  Handle both.
-    """
-    from ibis.expr.operations.sortkeys import SortKey as IbisSortKey
-
-    from .._xorq import SortKey as XorqSortKey, map_ibis
-
-    if IbisSortKey in map_ibis.registry:
-        return  # already patched
-
-    @map_ibis.register(IbisSortKey)
-    def _map_sort_key(val, kwargs=None):
-        # ibis 12 uses .arg, ibis 11 uses .expr
-        sort_expr = getattr(val, "arg", None) or getattr(val, "expr")
-        return XorqSortKey(
-            expr=map_ibis(sort_expr, None),
-            ascending=val.ascending,
-            nulls_first=val.nulls_first,
-        )
-
-
-def _ensure_xorq_table(table):
-    """Convert plain ibis Table to xorq-vendored ibis if possible.
-
-    This is the single boundary between user-supplied ibis tables and
-    BSL's internal xorq representation. ``SemanticModel`` calls it once
-    at construction so internal code paths can assume xorq tables when
-    the backend is supported, and a plain ibis fallback otherwise.
-
-    Falls back to returning the plain ibis table when the backend is not
-    supported by xorq (e.g. Databricks). Idempotent: calling it on a
-    xorq-vendored table is a cheap no-op.
-    """
-    _patch_xorq_sortkey_compat()
-    if "xorq.vendor.ibis" not in type(table).__module__:
-        try:
-            from .._xorq import from_ibis
-
-            return from_ibis(table)
-        except Exception:
-            # Backend isn't supported by xorq's map_ibis registry (e.g.
-            # Databricks). Fall back so plain-ibis paths can still execute.
-            logger.debug(
-                "from_ibis failed for %s; using plain ibis table",
-                type(table).__module__,
-                exc_info=True,
-            )
-            return table
-    return table
-
-
-def _rebind_to_backend(expr, target_backend):
-    """Rebind every ``DatabaseTable`` op in *expr* to *target_backend*.
-
-    Low-level primitive shared with ``serialization.reconstruct``.
-    No-op on plain ibis expressions or when xorq is unavailable for any
-    reason; callers must pass a xorq-vendored ``target_backend``.
-    """
-    try:
-        from .._xorq import relations as xorq_rel
-    except ImportError:
-        return expr
-
-    def _recreate(op, _kwargs, **overrides):
-        kwargs = dict(zip(op.__argnames__, op.__args__, strict=False))
-        if _kwargs:
-            kwargs.update(_kwargs)
-        kwargs.update(overrides)
-        return op.__recreate__(kwargs)
-
-    def replacer(op, _kwargs):
-        if isinstance(op, xorq_rel.DatabaseTable) and op.source is not target_backend:
-            return _recreate(op, _kwargs, source=target_backend)
-        if _kwargs:
-            return _recreate(op, _kwargs)
-        return op
-
-    return expr.op().replace(replacer).to_expr()
-
-
-def _rebind_to_canonical_backend(expr):
-    """Rebind divergent ``DatabaseTable`` backends in *expr* to share one.
-
-    ``from_ibis()`` creates a distinct ``Backend`` per call, so expressions
-    built by composing separately-converted tables contain multiple
-    backends. Picking the first ``DatabaseTable``'s source as canonical
-    and rebinding the rest eliminates "Multiple backends found" errors
-    at execution time.
-
-    No-op on plain ibis expressions (not xorq-vendored).
-    """
-    try:
-        from .._xorq import relations as xorq_rel, walk_nodes
-    except ImportError:
-        return expr
-
-    try:
-        db_tables = list(walk_nodes((xorq_rel.DatabaseTable,), expr))
-    except Exception:
-        # walk_nodes can't traverse plain ibis trees; treat as no-op.
-        logger.debug("walk_nodes failed on plain ibis expr", exc_info=True)
-        return expr
-
-    canonical = db_tables[0].source if db_tables else None
-    if canonical is None:
-        return expr
-
-    return _rebind_to_backend(expr, canonical)
+from ._xorq_compat import (  # noqa: E402
+    _ensure_xorq_table,
+    _patch_xorq_sortkey_compat,
+    _rebind_to_backend,
+    _rebind_to_canonical_backend,
+)
 
 
 def _to_untagged(source: Any) -> ir.Table:
@@ -601,50 +405,7 @@ def _classify_dependencies(
     }
 
 
-@frozen
-class _CallableWrapper:
-    """Hashable wrapper for Callable and Deferred.
-
-    Both raw callables (lambda) and user Deferred (_.foo) are not hashable
-    and cannot be stored in FrozenDict. This wrapper provides hashability
-    using identity-based hashing.
-    """
-
-    _fn: Any
-
-    def __call__(self, *args, **kwargs):
-        return self._fn(*args, **kwargs)
-
-    def __hash__(self):
-        # should this be dask.base.tokenize()?
-        return hash(id(self._fn))
-
-    @property
-    def unwrap(self):
-        return self._fn
-
-
-def _ensure_wrapped(fn: Any) -> _CallableWrapper:
-    """Wrap Callable or Deferred for hashability."""
-    return fn if isinstance(fn, _CallableWrapper) else _CallableWrapper(fn)
-
-
-def _infer_unnest(fn: Callable, table: Any) -> tuple[str, ...]:
-    """Infer required unnest operations from the table.
-
-    Examples:
-        to_semantic_table(tbl).with_measures(...) -> ()  # Session level
-        to_semantic_table(tbl).unnest("hits").with_measures(...) -> ("hits",)
-        unnested.unnest("product").with_measures(...) -> ("product",)
-    """
-    from ..expr import SemanticUnnest
-
-    if isinstance(table, SemanticUnnest):
-        op = table.op()
-        # SemanticUnnestOp always has column attribute
-        return (op.column,)
-
-    return ()
+from ._callable import _CallableWrapper, _ensure_wrapped, _infer_unnest  # noqa: E402
 
 
 def _extract_measure_metadata(
@@ -4396,453 +4157,18 @@ class SemanticIndexOp(Relation):
         return func(self, *args, **kwargs)
 
 
-def _find_root_model(node: Any) -> SemanticTableOp | None:
-    """Find root SemanticTableOp in the operation tree."""
-    cur = node
-    while cur is not None:
-        if isinstance(cur, SemanticTableOp):
-            return cur
-        parent = getattr(cur, "source", None)
-        cur = parent
-    return None
-
-
-def _find_all_root_models(node: Any) -> tuple[SemanticTableOp, ...]:
-    """Find all root SemanticTableOps in the operation tree (handles joins with multiple roots)."""
-    if isinstance(node, SemanticTableOp):
-        return [node]
-
-    roots = []
-
-    if hasattr(node, "left") and hasattr(node, "right"):
-        roots.extend(_find_all_root_models(node.left))
-        roots.extend(_find_all_root_models(node.right))
-    elif hasattr(node, "source") and node.source is not None:
-        roots.extend(_find_all_root_models(node.source))
-
-    return roots
-
-
-def _dimension_only_source_table(
-    keys: tuple[str, ...],
-    all_roots: Sequence[SemanticTableOp],
-    filters: tuple,
-) -> tuple[SemanticTableOp, list[str], tuple] | None:
-    """Check if a dimension-only query can be routed to a single source table.
-
-    When all requested dimension keys share a single table prefix and that
-    prefix maps to a root model whose dimensions cover every key, we can
-    bypass the join and query the dimension table directly.  This ensures
-    dimension members with no matching fact rows are still returned.
-
-    *filters* are the ``_CallableWrapper`` predicates collected between the
-    aggregate and the underlying join.  Filters whose column references all
-    belong to the target table are forwarded; if any filter references columns
-    outside the target table the shortcut is disabled.
-
-    Returns ``(root_op, unprefixed_keys, applicable_filters)`` or ``None``.
-    """
-    if not keys:
-        return None
-
-    prefixes: set[str] = set()
-    unprefixed: list[str] = []
-    for key in keys:
-        if "." not in key:
-            return None  # Non-prefixed key — can't determine source
-        prefix, name = key.split(".", 1)
-        prefixes.add(prefix)
-        unprefixed.append(name)
-
-    if len(prefixes) != 1:
-        return None  # Keys span multiple tables
-
-    target_prefix = next(iter(prefixes))
-
-    for root in all_roots:
-        if root.name == target_prefix:
-            root_dims = root.get_dimensions()
-            if all(k in root_dims for k in unprefixed):
-                # Validate that every filter only touches columns present
-                # on the target dimension table.  If any filter references
-                # columns from other tables we cannot use the shortcut.
-                if filters:
-                    tbl = _to_untagged(root)
-                    tbl_cols = frozenset(tbl.columns) | frozenset(root_dims)
-                    for flt in filters:
-                        fn = _unwrap(flt) if hasattr(flt, "unwrap") else flt
-                        extraction = _extract_columns_from_callable(fn, tbl)
-                        if extraction.extraction_failed:
-                            return None  # Can't determine — bail out
-                        if not extraction.columns <= tbl_cols:
-                            return None  # References columns outside target
-                return root, unprefixed, filters
-
-    return None
-
-
-def _build_join_depth_map(node: Any) -> dict[str, int]:
-    """Map each leaf table name to its actual ibis rname depth.
-
-    ``SemanticJoinOp.to_untagged`` calls ``_join_depth`` to determine the
-    rname suffix for each join level.  ``_join_depth`` counts the number
-    of ``SemanticJoinOp`` ancestors on the *left* spine.  The right child
-    at depth *d* gets ``rname = _rname_for_depth(d)``.
-
-    For nested subtrees on the right side of a join, ibis applies the
-    inner subtree's rname independently.  So ``aircraft_models`` at inner
-    depth 1 gets ``_right``, not ``_right3`` even if the outer depth is 3.
-
-    This function mirrors ``_join_depth`` logic: walk down the left spine,
-    recording the right child's depth at each level.  If the right child is
-    itself a join tree, recurse to get inner depths for its leaves.
-    """
-    depth_map: dict[str, int] = {}
-
-    def _record_leaf(n, depth: int):
-        """Record a leaf table at the given depth."""
-        if isinstance(n, SemanticTableOp):
-            name = n.name
-            if name and name not in depth_map:
-                depth_map[name] = depth
-
-    def _walk_join_spine(n):
-        """Walk the left spine of a join tree, recording depths."""
-        if not isinstance(n, SemanticJoinOp):
-            # Leftmost leaf: depth 0 (root, never renamed)
-            _record_leaf(n, 0)
-            return
-
-        depth = SemanticJoinOp._join_depth(n)
-        # The right child is joined at this depth
-        right = n.right
-        if isinstance(right, SemanticJoinOp):
-            # Right is a subtree — its leaves get inner depths
-            inner_map = _build_join_depth_map(right)
-            for tname, idepth in inner_map.items():
-                if tname not in depth_map:
-                    if idepth == 0:
-                        # Leftmost leaf of subtree sits at the outer depth
-                        # (it receives the outer rname suffix if conflicting)
-                        depth_map[tname] = depth
-                    else:
-                        # Inner leaves keep their inner depth (inner rname)
-                        depth_map[tname] = idepth
-        else:
-            _record_leaf(right, depth)
-
-        # Recurse down the left spine
-        _walk_join_spine(n.left)
-
-    _walk_join_spine(node)
-    return depth_map
-
-
-def _update_measure_refs_in_calc(expr, prefix_map: dict[str, str]):
-    """
-    Recursively update MeasureRef names in a calculated measure expression.
-
-    Args:
-        expr: A MeasureExpr (MeasureRef, AllOf, BinOp, MethodCall, or literal)
-        prefix_map: Mapping from old name to new prefixed name
-
-    Returns:
-        Updated expression with prefixed MeasureRef names
-    """
-    if isinstance(expr, MeasureRef):
-        # Update the measure reference name if it's in the map
-        new_name = prefix_map.get(expr.name, expr.name)
-        return MeasureRef(new_name)
-    elif isinstance(expr, AllOf):
-        # Update the inner MeasureRef
-        updated_ref = _update_measure_refs_in_calc(expr.ref, prefix_map)
-        return AllOf(updated_ref)
-    elif isinstance(expr, MethodCall):
-        updated_receiver = _update_measure_refs_in_calc(expr.receiver, prefix_map)
-        return MethodCall(
-            receiver=updated_receiver,
-            method=expr.method,
-            args=expr.args,
-            kwargs=expr.kwargs,
-        )
-    elif isinstance(expr, BinOp):
-        # Recursively update left and right
-        updated_left = _update_measure_refs_in_calc(expr.left, prefix_map)
-        updated_right = _update_measure_refs_in_calc(expr.right, prefix_map)
-        return BinOp(op=expr.op, left=updated_left, right=updated_right)
-    else:
-        # Literal number or other - return as-is
-        return expr
-
-
-def _extract_join_key_column_names(source: Relation) -> set[str]:
-    """
-    Extract column names that ibis will merge (coalesce) during joins.
-
-    Ibis only merges join-key columns when **both** sides of an equi-join share
-    the **same** column name (e.g., ``l.code == r.code``).  When names differ
-    (e.g., ``l.carrier == r.code``), the right column gets a ``_right`` suffix
-    instead.  We return only the intersection of left/right key names so that
-    ``_check_and_add_rename`` correctly detects columns that need renaming.
-
-    Args:
-        source: The relation to search for join operations
-
-    Returns:
-        Set of column names that ibis merges (same-name equi-join keys)
-    """
-    join_keys: set[str] = set()
-
-    def find_joins(node):
-        """Recursively find join operations and extract merged key columns."""
-        if isinstance(node, SemanticJoinOp) and node.on:
-            try:
-                left_expr = node.left.to_expr() if hasattr(node.left, "to_expr") else node.left
-                right_expr = node.right.to_expr() if hasattr(node.right, "to_expr") else node.right
-                result = _extract_join_key_columns(node.on, left_expr, right_expr)
-                if result.is_success():
-                    # ibis merges only same-name equi-join columns
-                    join_keys.update(result.left_columns & result.right_columns)
-            except (AttributeError, TypeError):
-                pass
-
-        if hasattr(node, "left") and isinstance(node.left, Relation):
-            find_joins(node.left)
-        if hasattr(node, "right") and isinstance(node.right, Relation):
-            find_joins(node.right)
-        if hasattr(node, "source") and isinstance(node.source, Relation):
-            find_joins(node.source)
-
-    find_joins(source)
-    return join_keys
-
-
-def _build_column_rename_map(
-    all_roots: Sequence[SemanticTable],
-    field_accessor: callable,
-    source: Relation | None = None,
-) -> dict[str, str]:
-    """
-    Build a mapping of dimension names to their renamed column names in joined tables.
-
-    When Ibis joins tables with duplicate column names, it renames columns from later
-    tables with '_right' suffix. However, columns used as join keys are merged and
-    NOT renamed, so we exclude them from the rename map.
-
-    Uses graph_utils for generic traversal and the returns library for safe handling.
-
-    Args:
-        all_roots: List of root semantic tables in join order
-        field_accessor: Function to get fields (dimensions) from a root
-        source: Optional source relation to extract join keys from
-
-    Returns:
-        Dict mapping dimension names like 'airports.city' to renamed columns like 'city_right'
-    """
-    # Build column index using graph_utils (returns Result)
-    from returns.result import Failure
-
-    from ..graph_utils import build_column_index_from_roots, extract_column_from_dimension
-
-    column_index_result = build_column_index_from_roots(all_roots)
-    if isinstance(column_index_result, Failure):
-        # If we can't build the index, return empty map (dimensions will use fallback behavior)
-        return {}
-
-    column_index = column_index_result.value_or({})
-
-    # Extract join key columns to exclude from renaming
-    join_keys = _extract_join_key_column_names(source) if source else set()
-
-    # Build a map from table name → actual ibis join depth by walking the
-    # join tree.  The flat index in all_roots does NOT equal ibis join depth
-    # for nested joins (e.g. aircraft → aircraft_models inside a flights
-    # join tree), so we must compute it from the tree structure.
-    join_depth_map: dict[str, int] = {}
-    if source is not None:
-        join_depth_map = _build_join_depth_map(source)
-
-    # Process dimensions and determine which need renamed columns
-    rename_map = {}
-
-    for idx, root in enumerate(all_roots):
-        if not root.name:
-            continue
-
-        fields_dict = field_accessor(root)
-        if not fields_dict:
-            continue
-
-        root_tbl = root.to_untagged()
-        # Use the actual join depth if available, otherwise fall back to table_idx
-        effective_depth = join_depth_map.get(root.name, idx)
-
-        for field_name, field_value in fields_dict.items():
-            # Extract column name using graph_utils (returns Maybe)
-            column_maybe = extract_column_from_dimension(field_value, root_tbl)
-
-            # Use Maybe pattern from returns library
-            column_maybe.bind_optional(
-                lambda base_column: _check_and_add_rename(  # noqa: B023
-                    rename_map=rename_map,
-                    base_column=base_column,
-                    prefixed_name=f"{root.name}.{field_name}",  # noqa: B023
-                    table_idx=idx,  # noqa: B023
-                    column_index=column_index,
-                    join_keys=join_keys,
-                    join_depth=effective_depth,  # noqa: B023
-                )
-            )
-
-    return rename_map
-
-
-def _check_and_add_rename(
-    rename_map: dict[str, str],
-    base_column: str,
-    prefixed_name: str,
-    table_idx: int,
-    column_index: dict[str, list[int]],
-    join_keys: set[str],
-    join_depth: int | None = None,
-) -> None:
-    """
-    Check if a column needs renaming and add to rename map if so.
-
-    ``table_idx`` is the flat index in ``all_roots`` used to detect
-    whether an earlier table has the same column.  ``join_depth`` is
-    the actual ibis join depth (from ``_build_join_depth_map``) used
-    to compute the ``_right`` / ``_right2`` / … suffix.
-
-    Args:
-        rename_map: Map to update with renames
-        base_column: The base column name
-        prefixed_name: The prefixed dimension name (e.g., 'airports.city')
-        table_idx: Flat index in all_roots (for conflict detection)
-        column_index: Index of column occurrences
-        join_keys: Set of column names used as join keys (these don't get renamed)
-        join_depth: Actual ibis join depth for suffix computation (defaults to table_idx)
-    """
-    # Skip columns that are join keys - they get merged, not renamed
-    if base_column in join_keys:
-        return
-
-    depth = join_depth if join_depth is not None else table_idx
-
-    if base_column in column_index:
-        tables_with_column = column_index[base_column]
-        # Check if any table before this one (in flat order) has the same column
-        earlier_tables = [t for t in tables_with_column if t < table_idx]
-        if earlier_tables:
-            suffix = "_right" if depth <= 1 else f"_right{depth}"
-            rename_map[prefixed_name] = f"{base_column}{suffix}"
-
-
-def _wrap_dimension_for_renamed_column(dimension: Dimension, renamed_column: str) -> Dimension:
-    """
-    Wrap a dimension to access a renamed column in a joined table.
-
-    Args:
-        dimension: The original dimension
-        renamed_column: The renamed column name (e.g., 'city_right')
-
-    Returns:
-        A new Dimension that accesses the renamed column
-    """
-
-    # Create a new callable that accesses the renamed column
-    def renamed_accessor(table: ir.Table) -> ir.Value:
-        return table[renamed_column]
-
-    # Return a new Dimension with the wrapped callable but same metadata
-    return Dimension(
-        expr=renamed_accessor,
-        description=dimension.description,
-        is_entity=dimension.is_entity,
-        is_time_dimension=dimension.is_time_dimension,
-        is_event_timestamp=dimension.is_event_timestamp,
-        smallest_time_grain=dimension.smallest_time_grain,
-        derived_dimensions=dimension.derived_dimensions,
-    )
-
-
-def _merge_fields_with_prefixing(
-    all_roots: Sequence[SemanticTable],
-    field_accessor: callable,
-    source: Relation | None = None,
-) -> FrozenDict[str, Any]:
-    """
-    Generic function to merge any type of fields (dimensions, measures) with prefixing.
-
-    Args:
-        all_roots: List of SemanticTable root models
-        field_accessor: Function that takes a root and returns the fields dict (e.g. lambda r: r.dimensions)
-        source: Optional source relation to extract join keys from for proper column renaming
-
-    Returns:
-        FrozenDict mapping field names (always prefixed with table name) to field values
-    """
-    if not all_roots:
-        return FrozenDict()
-
-    merged_fields = {}
-
-    is_calc_measures = False
-    is_dimensions = False
-    if all_roots:
-        sample_fields = field_accessor(all_roots[0])
-        if sample_fields:
-            from ..measure_scope import AllOf, BinOp, MeasureRef, MethodCall
-
-            first_val = next(iter(sample_fields.values()), None)
-            is_calc_measures = isinstance(
-                first_val,
-                MeasureRef | AllOf | BinOp | MethodCall | int | float,
-            )
-            is_dimensions = isinstance(first_val, Dimension)
-
-    # For dimensions, build a column rename map to handle Ibis join conflicts
-    column_rename_map = {}
-    if is_dimensions:
-        column_rename_map = _build_column_rename_map(all_roots, field_accessor, source)
-
-    for root in all_roots:
-        root_name = root.name
-        fields_dict = field_accessor(root)
-
-        if is_calc_measures and root_name:
-            base_map = (
-                {k: f"{root_name}.{k}" for k in root.get_measures()}
-                if hasattr(root, "get_measures")
-                else {}
-            )
-            calc_map = (
-                {k: f"{root_name}.{k}" for k in root.get_calculated_measures()}
-                if hasattr(root, "get_calculated_measures")
-                else {}
-            )
-            prefix_map = {**base_map, **calc_map}
-
-        for field_name, field_value in fields_dict.items():
-            if root_name:
-                # Always use prefixed name with . separator
-                prefixed_name = f"{root_name}.{field_name}"
-
-                # If it's a calculated measure, update internal MeasureRefs
-                if is_calc_measures:
-                    field_value = _update_measure_refs_in_calc(field_value, prefix_map)
-                # If it's a dimension that needs column renaming, wrap the callable
-                elif is_dimensions and prefixed_name in column_rename_map:
-                    field_value = _wrap_dimension_for_renamed_column(
-                        field_value, column_rename_map[prefixed_name]
-                    )
-
-                merged_fields[prefixed_name] = field_value
-            else:
-                # Fallback to original name if no root name
-                merged_fields[field_name] = field_value
-
-    return FrozenDict(merged_fields)
+from ._root_models import (  # noqa: E402
+    _build_column_rename_map,
+    _build_join_depth_map,
+    _check_and_add_rename,
+    _dimension_only_source_table,
+    _extract_join_key_column_names,
+    _find_all_root_models,
+    _find_root_model,
+    _merge_fields_with_prefixing,
+    _update_measure_refs_in_calc,
+    _wrap_dimension_for_renamed_column,
+)
 
 # Column-extraction classes & helpers were moved to ._column_extraction.
 # Re-imported here so internal callers in this module keep working.
